requirements.txt
Flask>=3.0.0
Flask-Cors>=4.0.0
Flask-Bcrypt>=1.0.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
PyJWT>=2.8.0
gunicorn>=22.0.0
pytest>=8.0.0
requests>=2.32.0
beautifulsoup4>=4.12.0
# Security-focused packages
cryptography>=45.0.6
Werkzeug>=3.0.0
